+++ server/src/main/java/org/allaymc/server/container/impl/StonecutterContainerImpl.java
@@ -2,6 +2,8 @@ package org.allaymc.server.container.impl;
 
 import org.allaymc.api.container.ContainerTypes;
 import org.allaymc.api.container.interfaces.StonecutterContainer;
+import org.allaymc.api.item.recipe.input.CraftingRecipeInput;
+import org.allaymc.api.item.recipe.input.RecipeInput;
 
 /**
  * @author IWareQ
@@ -9,5 +11,10 @@ import org.allaymc.api.container.interfaces.StonecutterContainer;
 public class StonecutterContainerImpl extends BlockContainerImpl implements StonecutterContainer {
     public StonecutterContainerImpl() {
         super(ContainerTypes.STONECUTTER);
+    }
+
+    @Override
+    public RecipeInput createRecipeInput() {
+        return new CraftingRecipeInput(getItemStack(INPUT_SLOT));
     }
 }